refactor(structured-outputs): separate claude and openai examples
